# Update Beam and Plasma to Pure SoA