It finally works now.

Performance on A100:
```
amr.n_cell = 2047 2047 2000
beam.num_particles = 100000000
elec.ppc = 4 4
ions.ppc = 4 4
```
New:
```
TinyProfiler total time across processes [min...avg...max]: 200.5 ... 200.5 ... 200.5

--------------------------------------------------------------------------------------------------
Name                                               NCalls  Excl. Min  Excl. Avg  Excl. Max   Max %
--------------------------------------------------------------------------------------------------
ExplicitDeposition()                                 4000      58.13      58.13      58.13  28.99% <---
AdvancePlasmaParticles()                             4000      57.11      57.11      57.11  28.48% <---
DepositCurrent_PlasmaParticleContainer()             4000      54.09      54.09      54.09  26.98% <---
hpmg::MultiGrid::solve1()                            2000      19.15      19.15      19.15   9.55%
AnyDST::Execute()                                   12000      5.006      5.006      5.006   2.50%
sortBeamParticlesByBox()                                2      2.179      2.179      2.179   1.09%
FFTPoissonSolverDirichlet::SolvePoissonEquation()    6000     0.8317     0.8317     0.8317   0.41%
Hipace::SolveOneSlice()                              2000     0.6687     0.6687     0.6687   0.33%
Fields::ShiftSlices()                                2000     0.5452     0.5452     0.5452   0.27%
AdvanceBeamParticlesSlice()                          2000     0.5168     0.5168     0.5168   0.26% <---
Hipace::InitializeSxSyWithBeam()                     2000     0.3997     0.3997     0.3997   0.20%
Fields::LinCombination()                             4000     0.3667     0.3667     0.3667   0.18%
PlasmaParticleContainer::InitParticles                  2       0.27       0.27       0.27   0.13%
DepositCurrentSlice_BeamParticleContainer()          4000     0.2205     0.2205     0.2205   0.11% <---
Fields::SolveExmByAndEypBx()                         2000     0.2008     0.2008     0.2008   0.10%
BeamParticleContainer::InitParticles()                  1     0.1952     0.1952     0.1952   0.10%
Fields::Multiply()                                   2000     0.1198     0.1198     0.1198   0.06%
```
Old:
```
TinyProfiler total time across processes [min...avg...max]: 202.9 ... 202.9 ... 202.9

--------------------------------------------------------------------------------------------------
Name                                               NCalls  Excl. Min  Excl. Avg  Excl. Max   Max %
--------------------------------------------------------------------------------------------------
AdvancePlasmaParticles()                             4000      59.46      59.46      59.46  29.31% <---
ExplicitDeposition()                                 4000      58.33      58.33      58.33  28.75% <---
DepositCurrent_PlasmaParticleContainer()             4000      54.66      54.66      54.66  26.94% <---
hpmg::MultiGrid::solve1()                            2000      19.19      19.19      19.19   9.46%
AnyDST::Execute()                                   12000      5.032      5.032      5.032   2.48%
sortBeamParticlesByBox()                                2       1.65       1.65       1.65   0.81%
FFTPoissonSolverDirichlet::SolvePoissonEquation()    6000      0.832      0.832      0.832   0.41%
Hipace::SolveOneSlice()                              2000     0.6697     0.6697     0.6697   0.33%
Fields::ShiftSlices()                                2000     0.5455     0.5455     0.5455   0.27%
Hipace::InitializeSxSyWithBeam()                     2000     0.3998     0.3998     0.3998   0.20%
Fields::LinCombination()                             4000     0.3665     0.3665     0.3665   0.18%
AdvanceBeamParticlesSlice()                          2000      0.297      0.297      0.297   0.15% <---
PlasmaParticleContainer::InitParticles                  2     0.2808     0.2808     0.2808   0.14%
Fields::SolveExmByAndEypBx()                         2000      0.203      0.203      0.203   0.10%
BeamParticleContainer::InitParticles()                  1     0.1796     0.1796     0.1796   0.09%
DepositCurrentSlice_BeamParticleContainer()          4000     0.1766     0.1766     0.1766   0.09% <---
```

Note: The beam uses a permutation array so that’s why it is slower.


- [ ] **Small enough** (< few 100s of lines), otherwise it should probably be split into smaller PRs
- [x] **Tested** (describe the tests in the PR description)
- [x] **Runs on GPU** (basic: the code compiles and run well with the new module)
- [ ] **Contains an automated test** (checksum and/or comparison with theory)
- [x] **Documented**: all elements (classes and their members, functions, namespaces, etc.) are documented
- [x] **Constified** (All that can be `const` is `const`)
- [x] **Code is clean** (no unwanted comments, )
- [x] **Style and code conventions** are respected at the bottom of https://github.com/Hi-PACE/hipace
- [x] **Proper label and GitHub project**, if applicable