@ax3l 2023-05-31
@AlexanderSinn :
@atmyers and I were wondering if the new split particle ID cause issues for you, running over the generation of 2billion particles on a single rank?