@AlexanderSinn 2023-05-31
For the plasma we have all the particles in the same ParticleTile so we are more limited by the index type (int/unsigned int). I already ran a simulation with almost 2^31 plasma particles in two containers (500 GB total) on CPU before PureSoA, much more wouldn’t work but also would be impractically slow. The Plasma id is only used to distinguish between valid and invalid particles, so it could be made simpler. Related: https://github.com/Hi-PACE/hipace/pull/963

For the beam however the id is used for particle tracking diagnostics and a lot of beam particles are cheap performance wise. Currently we initialize everything on one rank and one ParticleTile so we are still index limited and memory limited when first sorting/reordering per box, but this might get fixed and then we will be limited by id(). We already use 2^30 beam particles sometimes.