@MaxThevenet 2023-06-01
If this becomes a problem, would it be possible to change the AMReX behavior to use more bits for `id` and less for `cpu`, maybe as a compile-time option? For HiPACE++, we don't do much with `cpu`, and just a few bits would be sufficient. (we don't need it yet, this is just for information.)

Another question: IIRC there's a safeguard in AMReX that would abort if we exceed the range of possible IDs, right? I think I saw this in `NextID`. Otherwise we should put one in HiPACE++.